% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM.R
\name{GLM}
\alias{GLM}
\title{Generalized Linear Model}
\description{
This class specializes \link{ParamRegrModel} and is the abstract
base class for generalized linear model objects like \link{NormalGLM}. It
handles the (inverse) link function.
}
\section{Super class}{
\code{\link[gofreg:ParamRegrModel]{gofreg::ParamRegrModel}} -> \code{GLM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GLM-new}{\code{GLM$new()}}
\item \href{#method-GLM-clone}{\code{GLM$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="gofreg" data-topic="ParamRegrModel" data-id="F_yx"><a href='../../gofreg/html/ParamRegrModel.html#method-ParamRegrModel-F_yx'><code>gofreg::ParamRegrModel$F_yx()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gofreg" data-topic="ParamRegrModel" data-id="f_yx"><a href='../../gofreg/html/ParamRegrModel.html#method-ParamRegrModel-f_yx'><code>gofreg::ParamRegrModel$f_yx()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gofreg" data-topic="ParamRegrModel" data-id="fit"><a href='../../gofreg/html/ParamRegrModel.html#method-ParamRegrModel-fit'><code>gofreg::ParamRegrModel$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gofreg" data-topic="ParamRegrModel" data-id="get_params"><a href='../../gofreg/html/ParamRegrModel.html#method-ParamRegrModel-get_params'><code>gofreg::ParamRegrModel$get_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gofreg" data-topic="ParamRegrModel" data-id="mean_yx"><a href='../../gofreg/html/ParamRegrModel.html#method-ParamRegrModel-mean_yx'><code>gofreg::ParamRegrModel$mean_yx()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gofreg" data-topic="ParamRegrModel" data-id="sample_yx"><a href='../../gofreg/html/ParamRegrModel.html#method-ParamRegrModel-sample_yx'><code>gofreg::ParamRegrModel$sample_yx()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gofreg" data-topic="ParamRegrModel" data-id="set_params"><a href='../../gofreg/html/ParamRegrModel.html#method-ParamRegrModel-set_params'><code>gofreg::ParamRegrModel$set_params()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GLM-new"></a>}}
\if{latex}{\out{\hypertarget{method-GLM-new}{}}}
\subsection{Method \code{new()}}{
Initialize an object of class GLM.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GLM$new(
  linkinv = function(u) {
     return(u)
 }
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{linkinv}}{inverse link function, defaults to identity function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GLM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GLM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GLM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
