% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFTest.R
\name{GOFTest}
\alias{GOFTest}
\title{Goodness-of-fit test for parametric regressin}
\description{
This class implements functions to calculate the test statistic for the original data
as well as the statistics for bootstrap samples. It also offers the possibilitz to computate the
corresponding bootstrap p-value.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GOFTest-new}{\code{GOFTest$new()}}
\item \href{#method-GOFTest-get_stat_orig}{\code{GOFTest$get_stat_orig()}}
\item \href{#method-GOFTest-get_stats_boot}{\code{GOFTest$get_stats_boot()}}
\item \href{#method-GOFTest-get_pvalue}{\code{GOFTest$get_pvalue()}}
\item \href{#method-GOFTest-clone}{\code{GOFTest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GOFTest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GOFTest-new}{}}}
\subsection{Method \code{new()}}{
Initialize an object of class GOFTest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOFTest$new(x, y, model_fitted, test_stat, boot_type, nboot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector of covariates}

\item{\code{y}}{response variables}

\item{\code{model_fitted}}{object of class ParamRegrModel with defined default parameters}

\item{\code{test_stat}}{object of class TestStatistic}

\item{\code{boot_type}}{a specification for the type of resampling scheme: "rsmpl" if covariates shall resampled or "keep" if they shall be kept the same}

\item{\code{nboot}}{number of bootstrap iterations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GOFTest-get_stat_orig"></a>}}
\if{latex}{\out{\hypertarget{method-GOFTest-get_stat_orig}{}}}
\subsection{Method \code{get_stat_orig()}}{
Calculates the test statistic for the original data and model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOFTest$get_stat_orig()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class \code{TestStatistic}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GOFTest-get_stats_boot"></a>}}
\if{latex}{\out{\hypertarget{method-GOFTest-get_stats_boot}{}}}
\subsection{Method \code{get_stats_boot()}}{
Calculates the test statistics for the resampled data and corresponding models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOFTest$get_stats_boot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector of objects of class \code{TestStatistic}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GOFTest-get_pvalue"></a>}}
\if{latex}{\out{\hypertarget{method-GOFTest-get_pvalue}{}}}
\subsection{Method \code{get_pvalue()}}{
Calculates the bootstrap p-value for the given model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOFTest$get_pvalue()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
p-value for the null hypothesis that \code{y} was generated according to \code{model}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GOFTest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GOFTest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GOFTest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
