% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEP.R
\name{MEP}
\alias{MEP}
\title{Marked empirical process test statistic for a given GLM}
\description{
This class inherits from \link{TestStatistic} and implements a
function to calculate the test statistic (and x-y-values that can be used
to plot the underlying process).

The process underlying the test statistic is given in Dikta & Scheer
(2021) and defined by \deqn{\bar{R}^1_n(u) = \frac{1}{\sqrt{n}}
  \sum_{i=1}^n \left( Y_i - m(X_i, \hat{\beta}_n) \right) I_{\{\hat{\beta}_n
  X_i \le u\}}, \quad -\infty \le u \le \infty.}{(see formula given in paper).}
}
\examples{
# Create an example dataset
set.seed(123)
n <- 100
x <- rbind(runif(n), rbinom(n, 1, 0.5))
model <- NormalGLM$new()
y <- model$sample_yx(x, params=list(beta=c(2,3), sd=1))

# Fit the correct model
model$fit(x, y, params_init=list(beta=c(1,1), sd=3), inplace = TRUE)

# Print value of test statistic and plot corresponding process
ts <- MEP$new()
ts$calc_stat(x, y, model)
print(ts)
ggplot2::ggplot() + ts$geom_ts_proc()

# Fit a wrong model
model2 <- NormalGLM$new(linkinv = function(u) {u+10})
model2$fit(x, y, params_init=list(beta=c(1,1), sd=3), inplace = TRUE)

# Print value of test statistic and plot corresponding process
ts2 <- MEP$new()
ts2$calc_stat(x, y, model2)
print(ts2)
ggplot2::ggplot() + ts2$geom_ts_proc()
}
\section{Super class}{
\code{\link[gofreg:TestStatistic]{gofreg::TestStatistic}} -> \code{MEP}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MEP-calc_stat}{\code{MEP$calc_stat()}}
\item \href{#method-MEP-clone}{\code{MEP$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="gofreg" data-topic="TestStatistic" data-id="geom_ts_proc"><a href='../../gofreg/html/TestStatistic.html#method-TestStatistic-geom_ts_proc'><code>gofreg::TestStatistic$geom_ts_proc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gofreg" data-topic="TestStatistic" data-id="get_plot_xy"><a href='../../gofreg/html/TestStatistic.html#method-TestStatistic-get_plot_xy'><code>gofreg::TestStatistic$get_plot_xy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gofreg" data-topic="TestStatistic" data-id="get_value"><a href='../../gofreg/html/TestStatistic.html#method-TestStatistic-get_value'><code>gofreg::TestStatistic$get_value()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gofreg" data-topic="TestStatistic" data-id="print"><a href='../../gofreg/html/TestStatistic.html#method-TestStatistic-print'><code>gofreg::TestStatistic$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MEP-calc_stat"></a>}}
\if{latex}{\out{\hypertarget{method-MEP-calc_stat}{}}}
\subsection{Method \code{calc_stat()}}{
Calculate the value of the test statistic for given data
(x,y) and a model to test for.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MEP$calc_stat(x, y, model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector of covariates}

\item{\code{y}}{response variable}

\item{\code{model}}{\link{ParamRegrModel} to test for}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MEP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MEP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MEP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
