% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamResamplingScheme.R
\name{ParamResamplingScheme}
\alias{ParamResamplingScheme}
\title{Parametric Resampling Scheme for (X,Y) data}
\value{
\code{list()} with tags x and y containing the resampled data
}
\description{
This class inherits from \link{ResamplingScheme} and implements a
function to resample new data based on the given data and assumed model.

The covariates are kept the same (default) or resampled from \code{data$x} and
the response variables are drawn according to \code{model$sample_yx()}.
}
\examples{
# Create an example dataset
set.seed(123)
n <- 10
x <- rbind(runif(n), rbinom(n, 1, 0.5))
model <- NormalGLM$new()
params <- list(beta=c(2,3), sd=1)
y <- model$sample_yx(x, params=params)
data <- list(x = x, y = y)

# Fit the model to the data
model$fit(x, y, params_init=params, inplace=TRUE)

# Resample from the model given data with or without resampling x
rs <- ParamResamplingScheme$new()
rs$resample(data, model)
rs$resample(data, model, rsmpl_x=TRUE)
}
\section{Super class}{
\code{\link[gofreg:ResamplingScheme]{gofreg::ResamplingScheme}} -> \code{ParamResamplingScheme}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ParamResamplingScheme-resample}{\code{ParamResamplingScheme$resample()}}
\item \href{#method-ParamResamplingScheme-clone}{\code{ParamResamplingScheme$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamResamplingScheme-resample"></a>}}
\if{latex}{\out{\hypertarget{method-ParamResamplingScheme-resample}{}}}
\subsection{Method \code{resample()}}{
Generate a new, resampled dataset of the same shape as data
following the given model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamResamplingScheme$resample(data, model, rsmpl_x = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{list()} with tags x and y containing the original data}

\item{\code{model}}{\link{ParamRegrModel} to use for the resampling}

\item{\code{rsmpl_x}}{if TRUE covariates are resampled, if FALSE (default)
covariates are kept the same}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamResamplingScheme-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ParamResamplingScheme-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamResamplingScheme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
