% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingScheme.R
\name{ResamplingScheme}
\alias{ResamplingScheme}
\title{ResamplingScheme Class}
\value{
\code{list()} with the same tags as data containing the resampled
data
}
\description{
This is the abstract base class for resampling scheme objects
like \link{ParamResamplingScheme}.

Resampling schemes are built around the key method \code{resample()} which
generates new data based on the given data and assumed model.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResamplingScheme-resample}{\code{ResamplingScheme$resample()}}
\item \href{#method-ResamplingScheme-clone}{\code{ResamplingScheme$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingScheme-resample"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingScheme-resample}{}}}
\subsection{Method \code{resample()}}{
Generate a new, resampled dataset of the same shape as data
following the given model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingScheme$resample(data, model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{list()} containing the original data}

\item{\code{model}}{\link{ParamRegrModel} to use for the resampling}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingScheme-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingScheme-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingScheme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
