% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestStatistic.R
\name{TestStatistic}
\alias{TestStatistic}
\title{R6 Class representing a test statistic}
\description{
This class contains an abstract methods to calculate the test statistic (and x-y-values that can be used to plot it).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TestStatistic-get_value}{\code{TestStatistic$get_value()}}
\item \href{#method-TestStatistic-calc_stat}{\code{TestStatistic$calc_stat()}}
\item \href{#method-TestStatistic-get_plot_xy}{\code{TestStatistic$get_plot_xy()}}
\item \href{#method-TestStatistic-clone}{\code{TestStatistic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestStatistic-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-TestStatistic-get_value}{}}}
\subsection{Method \code{get_value()}}{
Returns the value of the test statistic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestStatistic$get_value()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
value of the test statistic
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestStatistic-calc_stat"></a>}}
\if{latex}{\out{\hypertarget{method-TestStatistic-calc_stat}{}}}
\subsection{Method \code{calc_stat()}}{
Calculate the value of the test statistic for given data (x,y) and a model to test for
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestStatistic$calc_stat(x, y, model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector of covariates}

\item{\code{y}}{response variable}

\item{\code{model}}{model of type ParamRegrModel to test for}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestStatistic-get_plot_xy"></a>}}
\if{latex}{\out{\hypertarget{method-TestStatistic-get_plot_xy}{}}}
\subsection{Method \code{get_plot_xy()}}{
Returns vectors of x and y that can be used to plot the process corresponding to the test statistic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestStatistic$get_plot_xy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with plot.x and plot.y
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestStatistic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TestStatistic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestStatistic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
