% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondKolmY.R
\name{CondKolmY}
\alias{CondKolmY}
\title{Conditional Kolmogorov test statistic for the marginal distribution of Y}
\description{
This class implements a function of \code{TestStatistic} to calculate the test statistic (and x-y-values that can be used to plot it).
}
\section{Super class}{
\code{\link[gofreg:TestStatistic]{gofreg::TestStatistic}} -> \code{CondKolmY}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CondKolmY-calc_stat}{\code{CondKolmY$calc_stat()}}
\item \href{#method-CondKolmY-clone}{\code{CondKolmY$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="gofreg" data-topic="TestStatistic" data-id="get_plot_xy"><a href='../../gofreg/html/TestStatistic.html#method-TestStatistic-get_plot_xy'><code>gofreg::TestStatistic$get_plot_xy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gofreg" data-topic="TestStatistic" data-id="get_value"><a href='../../gofreg/html/TestStatistic.html#method-TestStatistic-get_value'><code>gofreg::TestStatistic$get_value()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CondKolmY-calc_stat"></a>}}
\if{latex}{\out{\hypertarget{method-CondKolmY-calc_stat}{}}}
\subsection{Method \code{calc_stat()}}{
Calculate the value of the test statistic for given data (x,y) and a model to test for
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CondKolmY$calc_stat(x, y, model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector of covariates}

\item{\code{y}}{response variable}

\item{\code{model}}{model of type ParamRegrModel to test for}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CondKolmY-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CondKolmY-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CondKolmY$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
