% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodFunctions.R
\name{loglik_xzd}
\alias{loglik_xzd}
\title{Negative Log-likelihood Function for a Censored Parametric Regression Model}
\usage{
loglik_xzd(data, model, params)
}
\arguments{
\item{data}{\code{list()} with tags x, z and delta containing the data}

\item{model}{\link{ParamRegrModel} to use for the likelihood function}

\item{params}{vector with model parameters to compute likelihood function for}
}
\value{
Value of the negative log-likelihood function
}
\description{
The log-likelihood function for a parametric regression model
under random censorship with data (x,z,delta) is given by the sum of the
logarithm of the conditional density of Y given X=x evaluated at z if z was
uncensored or the logarithm of the conditional survival of Y given X=x
evaluated at z if z was censored.

This function is one option that can be used to fit a \link{ParamRegrModel}. It
returns the negative log-likelihood value in order for \code{optim()} to
maximize (instead of minimize).
}
