% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalGLM.R
\name{NormalGLM}
\alias{NormalGLM}
\title{Generalized linear model with normal distribution}
\description{
This class implements functions of \code{ParamRegrModel} that, for example, evaluate the conditional density and distribution functions.
}
\section{Super class}{
\code{\link[gofreg:ParamRegrModel]{gofreg::ParamRegrModel}} -> \code{NormalGLM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NormalGLM-new}{\code{NormalGLM$new()}}
\item \href{#method-NormalGLM-fit}{\code{NormalGLM$fit()}}
\item \href{#method-NormalGLM-f_yx}{\code{NormalGLM$f_yx()}}
\item \href{#method-NormalGLM-F_yx}{\code{NormalGLM$F_yx()}}
\item \href{#method-NormalGLM-sample_yx}{\code{NormalGLM$sample_yx()}}
\item \href{#method-NormalGLM-mean_yx}{\code{NormalGLM$mean_yx()}}
\item \href{#method-NormalGLM-clone}{\code{NormalGLM$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="gofreg" data-topic="ParamRegrModel" data-id="get_params"><a href='../../gofreg/html/ParamRegrModel.html#method-ParamRegrModel-get_params'><code>gofreg::ParamRegrModel$get_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gofreg" data-topic="ParamRegrModel" data-id="set_params"><a href='../../gofreg/html/ParamRegrModel.html#method-ParamRegrModel-set_params'><code>gofreg::ParamRegrModel$set_params()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalGLM-new"></a>}}
\if{latex}{\out{\hypertarget{method-NormalGLM-new}{}}}
\subsection{Method \code{new()}}{
Initialize an object of class NormalGLM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalGLM$new(linkinv)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{linkinv}}{inverse link function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalGLM-fit"></a>}}
\if{latex}{\out{\hypertarget{method-NormalGLM-fit}{}}}
\subsection{Method \code{fit()}}{
Calculates the maximum likelihood estimator for the model parameters based on given data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalGLM$fit(x, y, params_init = private$params, inplace = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector of covariates}

\item{\code{y}}{response variable}

\item{\code{params_init}}{initial value of the model parameters to use for the optimization}

\item{\code{inplace}}{logical; if TRUE, default model parameters are set accordingly and parameter estimator is not returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
MLE of the model parameters for the given data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalGLM-f_yx"></a>}}
\if{latex}{\out{\hypertarget{method-NormalGLM-f_yx}{}}}
\subsection{Method \code{f_yx()}}{
Evaluates the conditional density function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalGLM$f_yx(t, x, params = private$params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{value(s) at which the conditional density shall be evaluated}

\item{\code{x}}{vector of covariates}

\item{\code{params}}{use different model parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
value(s) of the conditional density function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalGLM-F_yx"></a>}}
\if{latex}{\out{\hypertarget{method-NormalGLM-F_yx}{}}}
\subsection{Method \code{F_yx()}}{
Evaluates the conditional distribution function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalGLM$F_yx(t, x, params = private$params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{value(s) at which the conditional distribution shall be evaluated}

\item{\code{x}}{vector of covariates}

\item{\code{params}}{use different model parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
value(s) of the conditional distribution function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalGLM-sample_yx"></a>}}
\if{latex}{\out{\hypertarget{method-NormalGLM-sample_yx}{}}}
\subsection{Method \code{sample_yx()}}{
Generates a new sample of response variables with the same conditional distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalGLM$sample_yx(x, params = private$params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector of covariates}

\item{\code{params}}{use different model parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of sampled response variables
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalGLM-mean_yx"></a>}}
\if{latex}{\out{\hypertarget{method-NormalGLM-mean_yx}{}}}
\subsection{Method \code{mean_yx()}}{
Evaluates the regression function or in other terms the expected value of Y given X=x
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalGLM$mean_yx(x, params = private$params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector of covariates}

\item{\code{params}}{use different model parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
value of the regression function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalGLM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NormalGLM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalGLM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
