% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamRegrModel.R
\name{ParamRegrModel}
\alias{ParamRegrModel}
\title{R6 Class representing a parametric regression model}
\description{
This class implements the maximum likelihood estimator for the model parameters
as well as abstract methods to, for example, evaluate the conditional density and distribution functions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ParamRegrModel-set_params}{\code{ParamRegrModel$set_params()}}
\item \href{#method-ParamRegrModel-get_params}{\code{ParamRegrModel$get_params()}}
\item \href{#method-ParamRegrModel-fit}{\code{ParamRegrModel$fit()}}
\item \href{#method-ParamRegrModel-f_yx}{\code{ParamRegrModel$f_yx()}}
\item \href{#method-ParamRegrModel-F_yx}{\code{ParamRegrModel$F_yx()}}
\item \href{#method-ParamRegrModel-sample_yx}{\code{ParamRegrModel$sample_yx()}}
\item \href{#method-ParamRegrModel-mean_yx}{\code{ParamRegrModel$mean_yx()}}
\item \href{#method-ParamRegrModel-clone}{\code{ParamRegrModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamRegrModel-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-ParamRegrModel-set_params}{}}}
\subsection{Method \code{set_params()}}{
Set the value of the model parameters used as default for the class functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamRegrModel$set_params(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{model parameters to use as default}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamRegrModel-get_params"></a>}}
\if{latex}{\out{\hypertarget{method-ParamRegrModel-get_params}{}}}
\subsection{Method \code{get_params()}}{
Returns the value of the model parameters used as default for the class functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamRegrModel$get_params()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
model parameters used as default
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamRegrModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-ParamRegrModel-fit}{}}}
\subsection{Method \code{fit()}}{
Calculates the maximum likelihood estimator for the model parameters based on given data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamRegrModel$fit(x, y, params_init = private$params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector of covariates}

\item{\code{y}}{response variable}

\item{\code{params_init}}{initial value of the model parameters to use for the optimization}

\item{\code{inplace}}{logical; if TRUE, default model parameters are set accordingly}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
MLE of the model parameters for the given data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamRegrModel-f_yx"></a>}}
\if{latex}{\out{\hypertarget{method-ParamRegrModel-f_yx}{}}}
\subsection{Method \code{f_yx()}}{
Evaluates the conditional density function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamRegrModel$f_yx(t, x, params = private$params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{value(s) at which the conditional density shall be evaluated}

\item{\code{x}}{vector of covariates}

\item{\code{params}}{use different model parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
value(s) of the conditional density function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamRegrModel-F_yx"></a>}}
\if{latex}{\out{\hypertarget{method-ParamRegrModel-F_yx}{}}}
\subsection{Method \code{F_yx()}}{
Evaluates the conditional distribution function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamRegrModel$F_yx(t, x, params = private$params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{value(s) at which the conditional distribution shall be evaluated}

\item{\code{x}}{vector of covariates}

\item{\code{params}}{use different model parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
value(s) of the conditional distribution function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamRegrModel-sample_yx"></a>}}
\if{latex}{\out{\hypertarget{method-ParamRegrModel-sample_yx}{}}}
\subsection{Method \code{sample_yx()}}{
Generates a new sample of response variables with the same conditional distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamRegrModel$sample_yx(x, params = private$params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector of covariates}

\item{\code{params}}{use different model parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of sampled response variables
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamRegrModel-mean_yx"></a>}}
\if{latex}{\out{\hypertarget{method-ParamRegrModel-mean_yx}{}}}
\subsection{Method \code{mean_yx()}}{
Evaluates the regression function or in other terms the expected value of Y given X=x
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamRegrModel$mean_yx(x, params = private$params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector of covariates}

\item{\code{params}}{use different model parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
value of the regression function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamRegrModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ParamRegrModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamRegrModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
